PDF interpreter - check for (illegal) non-dict XObjects when scanning spot colours
Bug #696288 "Error: /typecheck in --run-- writing psdcmyk file"

The PDF file contains a /Pages dictionary which includes a /Resources
dictionary, this is consulted if a page calls for a resource which isn't
defined in the page's own Resources dictionary.

Unfortunately the XObject dictionary defined in this resource contains a
reference to an object which is not an XObject.

When scanning for transparency we deal with this (and ignore it) but when
scanning for page spot colours, we do not, resulting in a typecheck error.

This commit adds checking the XObject type when scanning for page spot
colours, and ignores anything which isn't a dictionary.

No differences expected.