Fix multi-threaded rendering crash on Windows. Bug 696254.
The use of fseek(...SEEK_END)...ftell to determine the file size is not
thread safe because reading changes the current position. On Windows, the
ReadFile changes the position of the 'fd' attached to the stream and handle,
and the ftell uses 'lseek(fd, 0, SEEK_CUR)' to get the current position
which may have moved due to reads on other threads.

On unix, similar conditions can occur sharing the FILE * stream. Also on
unix, we must assume that without PREAD support, we cannot share the file
descriptor since reading requires "ftell..seek..read..seek" sequences that
are not thread safe without locking (that is not yet implemented, and may
impact performance).

Maintain the filesize while writing the file in the IFILE wrapper stucture
and use that for the file size instead of fseek..ftell to avoid the position
change that could happen on the clist file due to a read by a different
thread on that same file fd.

This was a rare problem because only cl_cache_read_init used the seek..tell
mechanism to get the filesize, and this only happens in a thread on the
first read (to either the bfile or the cfile) which is a narrow window.