Add proper check for gx_dc_type_data_devn color type in shading color linearity
When we added the gx_dc_type_data_devn to support the Device N planar devices,
we failed to update the linearity check for shadings when we draw to these
devices.   This resulted in the linearity check always failing which resulted
in extreme decomposition of the shading beyond what was needed.

Fixes bug 696290.  Note that a couple files like 442-01.ps will have some minor
shading artifacts visible that were not there before for the psdcmyk device.
I checked, and these same artifacts are visible in other CMYK devices (e.g. tiff32nc)
for this file.

These issues are different than what was addressed by this fix and should not
be considered a regression by this commit.  That said, we may want to open a bug
related to those issues.