Bug 695042; -dUsePDFX3Profile with -dNumRenderingThreads issues
When a PDF file has on Output Intent (OI) profile and we want
to use it as the device profile or the proof profile, it is
set for the device through a call from the interpreter in
zset_outputintent.   Unlike other profiles for the device,
this profile is not a device param that is set with get/put
params of a file name.  Instead the profile is inside the source
file that is being rendered.

When the threads are set up in setup_device_and_mem_for_thread
the icc profiles are set through put and get param calls. Since
the OI profile is not available in this manner, this fails.

This fix detects the presence of the OI profile during
setup_device_and_mem_for_thread and clones the profiles.
Doing a straight reference to the profile instead of cloning
is problematic if the CMM is not thread safe in its use
of the profile handles (which is true for LCMS), hence the
cloning process.

In addition, in this commit a non-ascii character was added to
the special name for the OI profile to help avoid issues with
an actual profile being named with our special key word.