PDF interpreter - handle broken type 1 fonts with no Private dict earlier
Bug #696306 "Incomplete rendering of PDF file"

The embedded type 1 font 'Consolas' is broken. The eexec encrypted
portion of the font suddenly ends during a glyph description. This means
that the Private dictionary is never added to the font dictionary but is
instead left on the operand stack. Where the PDF interpreter removes it
in order to avoid other errors.

The error actually occurs inside .buildfont1, but despite efforts it
proved impossible to deal with the error at that point. This was due to
all the code already in place to deal with other errors while handling
broken PDF files. It wasn't possible to address this particular problem
without breaking other (admittedly invalid) PDF files.

This commit checks the font dictionary for the presence of a Private
dictionary before returning from readtype1 and if the Private dict is
not present it instead returns a null object, which signals the calling
code to attempt to load the font from another source.

No differences expected