Replacement fix for commit 487ed6
Bug #696306 was originally fixed in commit 487ed6. After talking about
it with Chris, it became clearer why I was unable to intercept the
broken font where I wanted. gs_type1.ps contains a redefinition of
.buildfont1, and this redefinition does not properly preserve the
stack in the event of an error.

If this was a simple procedure that would be understandable, but the
redefinition is of a C operator, so not preserving the stack is actually
wrong.

The commit here corrects that problem, which enables me to intercept the
error much later in the page rendering and do a considerably better job
of correcting the problem. This now renders as per Marcos' Acrobat
screenshot (but unlike mine).

This fix should be used in preference.