Bug 696305: Revamp scan_contour
As part of the investigation into bug 696305, rejig the functions
around 'scan_contour' for clarity.

Firstly, we pull the usage of contour_cursor into scan_contour.
Rather than having the caller set part of the structure up, and
then pass it in, have the caller pass in the required parts and
let all of the structure handling be done within scan_contour.
This improves the modularity of this code at least.

Secondly, rather than having the contour_cursor structure contain
a pointer to a gx_flattened_iterator allocated separately on the
stack, have it be part of the structure. This saves lots of
dereferencing (which might help with repeated reads due to C's
pointer aliasing) and is clearer to read.

Thirdly, we break the if's into more deeply nested things. This
results in more indented code, but it avoids us having to retest
the same variables several times, and may give us some improvement
in runtimes.

Firstly, add some comments. This is a break with tradition in
this code. These comments describe my understanding (however
limited) it may be of what the code is doing.