Bug 696305: Ensure subdivided lines are not dropped
While walking the contour in scan_contour, we use
gx_flattened_iterator to cope with flattening curves.
Each time we call gx_flattened_iterator__next we get the
next section of the line out.

To avoid overflows in huge lines/curves the iterator splits
segments into 4. This was showing up a problem in this code
whereby a vertical line from 0x80xxxxxx to 0x7fxxxxxx was
being split into 4. 2 of these sub-divided lines crossed the
region of interest, but only the first of them was being
considered for addition into the active line list.

The fix here is simply to ensure that we continue to loop
through the iterator for the curve if we fail to add anything
from it.

It is possible that there are further optimisations possible
here (see the FIXMEs in the code), but the changes I've made
should be safe and minimally invasive.

Many thanks to Chris for doing the legwork in tracking the
cause down!