Fix Bug 696324 SMask None handled incorrectly by clist writing.
The SMask None special case that is sent using gssmask in the PDF
interpreter when it needs to make sure there is no SMask in place
needs special handling w.r.t. the clist cropping logic since we do
want to write it, but there won't be a endtransparencymask to
perform the pop of the cropping stack.