PDF Interpreter - Ignore empty /Kids arrays in AcroForm fields
Bug #696342 "PDF Annotation Error: /rangecheck in --run--"

The PDF file has an AcroForm with a field where the /Kids array is [ ]
This previously caused an error.

Adopting the patch from Martin McNabb with a very slight tweak to issue
a warning if -dQUIET isn't set.

Bizarrely this does exhibit a progression with Bug694429.pdf on the
cluster, even though this code isn't executed......

Its a progression so I'm not going to complain.