Bug 694238: Fix segfault in error during PatternType 1
If the PaintProc of a Type 1 pattern triggers an error after having done one
or more gsaves, we'll try to retrieve the pattern instance from the wrong
graphics state when we attempt the final cleanup.

To address this, store a reference to the pattern instance on the exec stack
which a) guarantees we get the correct pattern instance during cleanup,
and b) allows us to roll back the graphics state stack to the correct point.

No cluster differences.