Bug 694238: init several gs_glyph_info_t structs
In various places we were calling a font's glyph_info which may, or may not
fully fill in the gs_glyph_info_t passed to it, then using the results,
regardless of whether the specific value had been set. Initialize to zeros
for at least consistent results.

No cluster differences.