Bug 696345: .nativeFontmap when TTF name table is invalid
When building the .nativeFontmap (using fonts retrieved from a platform
specific API - i.e. fontconfig), if we can't read the font name from the
font file, we fall back to using the font name as reported by the API.

But, the code failed to take into account that the operand stack has different
depth depending on whether we were able to read the name from the file or not.

This commit handles that by using a counttomark rather than hard coded stack
depth.

No cluster differences.