Bug 694237: Handle missing/incomplete TTF glyph lengths
In the case of a broken TTF based CIDFont passing through ps2write (or one of
the pdfwrite paths that requires glyphs to be rendered), we can end up with
a partially complete font structure - in particular, the glyph lengths table
may be incomplete or missing.

We have a couple of fallback options available to get the length of a glyph:
the first is to retrieve the offset of the data for *next* glyph index and
the difference between the two offsets is the length of the glyph of interest.
If that fails (particularly if we are already processing the last available
glyph), we can use the offset to the end of the sfnt data - since the glyph
table is invariably the last table in the sfnt stream.

No cluster differences.