graphics library - if pattern x or y size is 0, don't estimate tile size
Inspired by Bug #696351

The bug report in Bug #696351 is ridiculously incomplete and the reporter
seems determined not to provide any real assistance.

However, by some logic and experimentation it did prove possible to
(eventually) reproduce the problem and trace through the code from
pattern creation. (run with a very low resolution, -r10 did it for me)

The crash is caused by attempting to estimate the size of a pattern
bitmap tile, when the pattern has a size of 0 in the y direction. This
is a legitimate value, we simply drop the pattern in this case. Since
we aren't going to render anything, the tile will have a size of 0 so
we can easily short-circuit all this calculation by testing for the tile
being 0 in either the x or y direction and simply returning 0.

No differences expected.