Bug 696316: Enforce memory alignment for libpng
libpng's inclusion of it's jmp buffer at the beginning of it's private context
requires the allocation to be 16 byte aligned on Win64 (because the jmp
buffer must be 16 byte aligned).

In the libpng callback functions for memory allocationg and freeing enforce
that alignment.

Also, add those callbacks (including the alignment enforcement) to the gdevpng.c
output device.

(Code mostly supplied by Robin!).

No cluster differences.