Hash CIE spaces to detect matching, cached, ICC profiles
Bug #696355 "Create unique ID for CIE color spaces"

Creating an ICC profile for a PostScript colour space is a performance
hit. Especially (I believe) for CIE spaces. We maintain a cache of ICC
profiles that have been created, but we need a way to identify if a
given, cached, profile matches a newly selected colour space.

There is code already in place for this, but missing the generation of
a unique ID for a space, so that we can find a matching profile, if we
have one cached.

This commit uses the existing MD5 machinery to create a hash from the
PostScript array defining a CIEBased colour space. We then use that hash
as the ID for the space, and check to see if we already have a cached
ICC profile with a matching ID.

This should improve performance on files using CIEBased colour spaces,
especially if they do 'gsave [/CIEBased <<...>>] setcolorspace grestore'
operations, as we will only need to create the ICC profile once.

I've created and manually checked example CIEBasedA, CIEBasedABC,
CIEBasedDEF and CIEBasedDEFG files to see that two identical spaces are
correctly identified as the same and that spaces with even very tiny
differences are correctly identified as different.

No differences expected in cluster test.