PDF Interpreter - Handle incorrect /Size in Xref streams
Bug #696365 "Error reading PDF file"

The PDF file is, as usual, invalid. It uses Xref streams and is a
Linearized file, meaning that there are two Xrefs, the first one being
for page 1 (caused by Linearization).

This xref looks like ths:

<<
/Size 12
/Root 30 0 R
/Prev 15073
/Info 28 0 R
/ID[<B6AEC95A19F1E4391AFF6AF538489730><B6AEC95A19F1E4391AFF6AF538489730>]
/Type/XRef/Index[29 39]/W[1 4 1]/Filter /FlateDecode /Length 112
>>

Notice that the Size (the number of entries in the xref) is given as 12
but the Index (starting index + number of entries) gives the size of the
table as 39, starting at index 29.

We were believing the Size, which meant that we created a Xref table
which was too small to hold the actual number of entries.

This commit checks both 'starting index + number of entries in the Index'
and 'starting index + Size' and uses the larger.

NB the xref also contains an entry with an invalid offset, but this does
not seem to cause a problem, presumably the entry is never actually used