Improve documentation and fixup toolbin/halftone tools
The thresh_remap (previously referred to as linearize_threshold) was
never in the package, so add it in a separate directory, and fix the
reference to it in gen_ordered.c

Add README files for the upper level, for thresh_remap and for the
gen_stochastic directories, so people don't have to read the code to
find the parameters.

Fix gen_stochastic minimum dot logic to actually work, although only
the 2x2 (-m5) parameter was tested. Also change the defaults for the
tolerance, -t, to 10 (1%) and exponential factor affecting choice, -p,
to 2.5 both of which tighten the selection to improve the quality (at
least in my opinion).