Fix the gs library context's allocator use.
The gs library context now uses the allocator with which it was
initialized for all memory operations.  Previously it used different
allocators, resulting in mismatched alloc/free pairs.  For example, the
languages passed in a memory pointer to the heap allocator upon
initialization of the context and then passed in a memory pointer to chunk
allocator when the library context is shut down.