Bug 695903: Augment fontmap syntax to allow TTCs to work 'properly'.
Support dictionary entries in Fontmap, so a fontmap file can contain a map to
specific fonts in a TTC.

The romfs inlines the fontmap definitions, and ends up parsing them before
we've bootstrapped the Level 2 world properly so the '<<...>>' dictionary
syntax not only is not defined but also cannot even be parsed.

So, temporarily enable Level 2 parsing, and include local definitions of the
names '<<' and '>>' to work as you'd expect.