EPSFitPage - fix some kinds of rotation
Bug #696128 "Rendering an EPS to a PNG with -dEPSFitPage and -gWxH sometimes results in blank render"

There was a typo in commit d59e1feb9545b399027907cb2d1a6855c524e0b4 which
prevented proper rotation in some cases. Note that in this particular
case we do *NOT* want to rotate.