pdfwrite - fix array dta source mesh shadings
Bug #696433 "Indeterminism with Bug695847b.ps and the pdfwrite device"

The code for emitting a mesh shading in a PDF given floating point
input (array based vertex data, at least) calcul;ates the number of
colour samples by multiplying the number of componentsd in the colour
space by the number of vertices.

However, when applying the max/min clamping from the ranges array in the
colour space, it used the current colour sample index, which obviously
could be much larger than the number of components in the colour space.

This caused us to run off the end of the pranges array and use
uninitialised data to set the colour sample, leading to non-deterministic
(and indeed incorrect) output.

We now pass in the number of components in the colour space and use the
modulus of that with the colour value index to index the pranges array.