Bug 695771: multithread/bgprint render may not be complete device shutdown
In some cleanup code, I assumed that by the time that code was called, rendering
would be complete - specifcally, background rendering would be complete.

This may be true when background rendering is one thread, but is definitely not
guaranteed when we have bgprint *and* multithreaded rendering.

Rather nicely, accounting for this actually makes the code tidier.

Also, add a check that we successfully created a clist IFILE object before
storing a value in it.