Bug 696441: handle numcopies > 1 in bgprint mode.
When numcopies is greater than 1, the device closes and closes and opens
the output file(s) for each copy. In the case of background printing, the
main device opens the output file, and generally expects to close the same
file, but with num_copies > 1 that is not the case.

When shutting down a bgprint "worker" device, copy the file pointer from
the worker device to the main device, so the main device retains a valid
output file pointer.