PDF interpreter - don't store ICCBased colour spaces in resolved dicts
Bug #696439 "unable to extract page one special page"

The page in question has nearly 2000 images, most in ICCbased colour
spaces, and large profiles (around 2MB or more). Because of the way we
dereference objects, this led to the ICCbased DataSource (the 2MB+)
being stored in the colour space, which was stored in the image
dictionaries, which were stored in the Page's Resources dicttionary.

The Page's Resources dictionary is not freed until the page is complete
which meant we were storing an awful lot of data. When we exceeded 2GB
(on the 32 bit build) we ran out of memory and mostly stopped rendering
anything further.

This commit alters the image (and shading, which has the same problem)
handling so that if the colour space is an array object, we make a copy
of the dictionary and use that. This leaves the original definition
unresolved in the Page's Resources dictionary, if we should need to use
the object again then we will have the (slight) overhead of needing to
rerun it, but at least we won't be storing gobs of mostly useless data
in the general case.

No differences expected.