Bug 696446: only use const strings for param list keys
The original fix replaced uses of strdup() (which caused a memory leak) with
a dynamically allocated temporary string which was freed at the end of
cups_get/put_params().

Unfortunately, as param lists simply take a reference to the key string, this
meant that the contents of the string were overwritten, and indeterminate
after being freed.

Switch to using lists of static const strings.

And add a stern warning to gsparams.h on the subject.

And squash a couple of compiler warnings in the cups code.