PDF interpreter - fix OutputIntent for device when no Trailer located in PDF (broken)
Bug #696447 "Error message lost with x11 device"

Despite the comment, I believe the X11 device behaviour is correct. We
generally try not to throw PostScript errors when dealing with PDF files,
even badly broken PDF files.

In this case the PDF is so broken we are unable to repair it (Acrobat
can't either), which means we have no Trailer either found or created.
However, we must continue with our processing loop and that includes
processing OutputIntents, if the device supports them.

The X11 device (and the Windows display device amongst others) do *NOT*
support an OutputIntent (its not file based so there's nowhere to put it)
but the ppmraw device, erroneously I believe, does claim to be able to
write an OutputIntent.

This led us into code in writeoutputintents whcih attempted to use the
Trailer without checking if it was present, and so throwing an error.

This commit simply checks if we have a Trailer, and doesn't attempt to
write the OutputIntent if we don't (because the PDF is unusable anyway)


This makes the devices behave the same, even though its not exactly
the fix Marcos wanted.