Bug 695576: tidy up xpsprint dependency confusion
Promote the pushing of jobs into the XPS print queue to be a "platform"
feature - i.e. make it a required "gp_" call, with a real one for Windows
(when supporting libs are available) and a dummy for when an XPS print queue
is not available.

This eases the confusion of dependencies and removes the need for platform
specific conditionally compiled code in the xpswrite device code.