JPEG/DCT dependency fix
Because we have common code to handle certain aspects of JPEG/DCT encoding and
decoding, we cannot include one without the other.

Add a "top level" sdct.dev which includes both sdcte.dev and sdctd.dev and make
that part of the core graphics lib features.

As our code and the libjpeg code both have considerable common code between
encoding and decoding, the inpact on the executable size is negligible.