PDF interpreter - handle (invalid) PS CMap streams
Bug #696449 "Incomplete rendering of PDF file"

The file includes a CIDFont which uses a custom CMap, which is incorrectly
defined. CMaps in PDF files are specified to be a CMap stream, that is a
dictionary declaration followed by a stream of data. The dictionary
should contain the CMApName, the CIDSystemInfo and WMode.

The supplied file has a stream which is simply a PostScript CMap, it
does not include any of the required definitions, and is, as far as I can
see, therefore invalid.

Nevertheless 'Acrobat can open it', so I've modified the type 0 font
and more specifically the CMap interpretation, to read a PostScript
CMap. This is not as easy as it may seem, and there may still be
problems with this code, because PostScript CMap files are normally
(unsurprisingly) run in a PostScript interpreter. Trying to handle them
without executing them in a PostScript interpreter is tricky. We can't
simply execute the PostScript (which would be ideal) because we need to
alter the CMAP Type.

This commit also contains a minor change to make Annotation processing
check the PDFSTOPONERROR flag when running annotations, and if it is set
cease processing, instead of ignoring the error.