PDF interpreter - improved recovery encountering broken fonts
Bug #696306 "Incomplete rendering of PDF file"

The original commit to fix this 487ed6d3b5fabbe21c23da288fbf020f49a28fae
did resolve the problem, however it was replaced with a better (better
in terms of correct PostScript error handling)
243614398b7bf3e8c4d080de7f8bbcb7436472cf

Unfortunately I only tested that with a decompressed file from MuPDF,
rather than the original file, and this failed with an additional
problem in the original file.

The file has a font stream which is ASCIIHex encoded, but contains,
at the end, some non-ASCIIHex characters (specifically a Q) which
causes an error in the ASCIIHexDecode before we even try to execute
the font (which is also damaged).

The commit to fix this fixed the handling of the broken font, but did
not work correctly with the invalid ASCIIHex stream. The problem was
that the font code returns a null object (instead of a font dictionary)
but does not signal an error. This means that the code for recovering
the operand and dictionary stacks did not execute, leaving the dict
stack in an incorrect state.

This code copies and executes the recovery code so that, even if there
is no error, we still clean up.

In addition, PDFSTOPONERROR is now honoured in a couple more places when
we encounter font problems.

Finally I haver altered pdfformaterror to put its output on stdout
instead of stderr, because this makes debugging much easier (the
warnings are now interleaved into the PDFDEBUG output for instance).

No differences expected