PDF Interpreter - improve performance on files with many xref sections
Bug #696454 "Regression: Performance decrease starting with 002cd5262ccb71010473abfb9069e1fb39f36f12"

The file is very large, and has an extremely large (~700,000 entries)
xref table with a very large number of sections in it. Normally we
allocate our internal xref storage as we parse the xref table, each time
we find a new section we increase the size of an array by the number
of entries in the section.

If we have a lot of sections, we end up spending a lot of time in the
memory management, allocating, copying and freeing arrays. This was less
of a problem when we used a 2D array because each array had a maximum of
65535 entries. However now that we use a single array this can be time
consuming.

The first part of this commit searches the trailer dictionaries (stored
after the xref) looking for the /Size entry. If we find one we use that
to set the initial size of the xref array. This saves us having to resize
the array as the Size is large enough to contain all entries.

However, this file *also* has an error in its /Size entry. It declares
the /Size as 88693, whereas in fact the xref contains nearly 700,000
entries. Our code still worked, but it was again spending a great deal
of time in memory management.

So the second part of this commit detects the fact that the declared
/Size is incorrect. When this happens, instead of allocating just
enough new entries in the array for the section we have discovered we
allocate 65534 entries, which makes it more likely that subsequent
sections will not need to increase the size again.

This is slightly wasteful of memory as we allocate storage we will never
use, but since the file is invalid it seems better to do this than
spend a lot of time minimising memory usage.

In my cluster tests this seems to give a small performance improvement

No other differences expected.