Fix for Bug 695074 x11 device encode color
The X11 device can be set up for different bit depths including
indexed like colors.  It could fill its palette when we have
a small number of possible colors and fail to encode.  Really the
device should not fail the encoding but always return the best
color that it can.  It should not rely upon gs to do any halftoning
to provide possibly see if another color could be encoded.

With this fix, we try to use a color that is in the existing palette,
adding a color if it is not available and we have space.  If we
no longer have space we now force the use of the closest color in
the table.