Bug 696142: Sanitise tiffsep separation names.
The existing code 'copies' separation names for printing/using in
filenames in a very naive way. It essentially blanks out any
characters unsupported in the filesystem with (or '%') with '_'.
Top bit set chars are let through unchanged, which causes
confusion as we now assume that filenames are held internally
in utf-8 format.

The code is updated here to output such filenames with simple
%02x style escaping. All characters unsupported in the filesystem
(or '%', or backspace, or top bit set chars) are escaped.

This leaves us utf-8 safe strings.

The only 'interesting' aspect here is that we have to check for
whether we need to put '%%' instead of '%' because we might be
printf-style processing the string later on.