PDF interpreter - reword warnings
Marcos has complained on customer's behalf on a number of occasions
that we don't use the word 'error' very much when fixing broken PDF
files leading to unrealistic expectations from the customer.

This commit breaks the errors and warnings up into two distinct groups;

Warnings are now uncommon and are reserved for situations where it seems
unlikely to me that the problem will result in a rendering error. For
example, any problems processing ToUnicode CMaps are now treated as
warnings since these are not used in rendering.

Anything which I feel has a reasonable chance of indicating a problem
with the PDF file sufficiently serious to result in a difference between
our output and Acrobat is now flagged as an error.

Each message should explicitly use the word error or warning as
appropriate, to avoid any confusion.

The end of job notifications differ slightly between errors and
warnings (errors state again that the output may be incorrect). This
potentially allows us to take additional different action in the future
as errors and warnings are tracked separately (NB errors trump
warnings, if you get an error, you get he error EOJ message)

These notifications now also obey QUIET, so they won't be show if this is
set.