pdfwrite - permit creation of PDF/A-3 files, disable file embedding if PDFA < 2
Bug #696473 "feature request: PDF/A-3 support"

There is actually very little practical difference between a PDF/A-2
and a PDF/A-3 file, PDF/A-3 permits embedding of arbitrary files
while PDF/A-2 does not.

Prior to commit a91d2576df0e60f6e691a3bd967b51109ae41f22 which
added support for the EMBED pdfmark there was therefore no point in
PDF/A-3 output.

This commit allows the -dPDFA switch to be turned up to 3 and creates
appropriate metadata. It also disables the EMBED pdfmark when creating
PDF/A-1 and emits a warning for PDF/A-2 that we don't validate the
content (PDF/A-2 only permits inclusion of PDF/A-1 and PDF/A-2). The
precise behaviour depends on the setting of PDFACompatibilityPolicy.

I can't claim this is well tested.