pdfwrite - If downsampling filter initialisation fails, don't downsample
If the image downsampling filter failed during initialisation we still
carried on and set up to downsample (we ignored the error return). This
would later result in a further error, and we would try to fall back
using the default image code. However, that would again try to set up
the downsampling filter, and would again fail, resulting in the image
being dropped from the output.

Here we check the return code from image initialisation, and if it fails
we emit a warnign and stop trying to do downsampling.