Documentation - sterner words on use of PDFSETTINGS
Many people use the 'PDFSETTINGS' switch without, aopparently, comprehending
that this will involve some alteration to the input. In addition there
seems to be a belief that '/prepress' will produce the best (ie closest
to the original) output, which is absolutely not the case. The best
quality is obtained by leaving pdfwrite alone.

I doubt it will make any difference, since so many people use a cargo cult
approach to Ghostscript command line parameters, but at least it is now
in the documentation.