pdfwrite - implement Metadata pdfmark and a non-standard extension metadata pdfmark
Bug #696472 "pdfwrite lacks support for /Metadata pdfmark"

There are several issues tackled in this single commit:

1) A previous commit left it impossible to use the EMBED pdfmark unless
producing PDF/A, fixed.

2) The Metadata pdfmark is now implemented. This allows the user to
specify an XMP stream which will be written to the Catalog of the PDF
file. The pdfwrite device will not overwrite this, or add any further
Metadata, it is the user's responsibility to get the XMP data correct.

3) A new pdfmark 'Ext_Metadata' has bee defined. This takes a string
parameter which contains XML to be add to the XMP normally created by
pdfwrite. Again ths is up to the user to specify correctly.

In no case does pdfwrite attempt to validate the data, though it will
issue warnings that PDF/A and PDF/X files may not be valid.

The additional pdfmark is to support creation of the ZugFERD electronic
invoice (http://www.ferd-net.de/front_content.php?idcat=231&changelang=4)
and potentially other similar formats in the future.

These new features are essentially untested.