Fix potential valgrind warning.
We don't have an example file for this, but this was spotted
during investigations into Bug 693784.

We can overrun by 1 pixel for odd length rows. When we overrun
we read the lower 4 bits of a byte, and those may be undefined.

Simply make PACIFY_VALGRIND blank these bits before we start.