Bug 697822: clist fix
When saving/restoring the clist state around the saved pages
processing, ensure that the file handling is correct. Leaving
the old filenames in play in particular is a bad thing, as the
shared fdesc stuff gets confused by this.

This commit reworks that so that clist files aren't closed/reopened
as much.