Bug 649971: GS Stroking fix
'Fat' strokes for curves of high curvature can go wrong due to the
winding for the 'underjoin' below curves being in the wrong sense.

This commit introduces a tweaked version of the code that solves this
at the cost of more complex paths. Sadly the cost for this seems to
be a huge slowdown, so the new code is disabled for now unless we
build with SLOWER_BUT_MORE_ACCURATE_STROKING defined.