Squash warning: switch(rop)
Ensure that rop3_T and rop3_S are defined as enum values as well
as preprocessor definitions.

They must be preprocessor definitions as otherwise the rop templating
code does not work. If they are not enum values, we get warned about
switch(rop) testing for values not defined in the enums.