Squash warnings: ICC code.
Various bits of the ICC code collects return values only to not
check them. Fix that here.

In some cases this is because functions have a 'void' return type
so there isn't a way to return an error. Change those return types
and ensure that all callers check and propogate the errors.