Squash warnings: MSVC ones.
Tweak the code to avoid the warnings seen in the MSVC debug build of
gs.

Mostly adding a few casts to make type changes explicit (and hence
avoid the "casting from int to double might lose data" etc warnings).

Reorder the headers in a couple of places to avoid offsetof being
redefined in a system header warnings.