Bug 696480: add a missing 'pop' for stack cleanup.
When attempting to create all viable glyph name mappings in a Type 1 font,
as a last ditch, we try to parse a unicode code point out of a string containing
a'uniXXXX' type glyph name (where 'XXXX' is the code point). If that parsing
fails, we were leaving the string on the stack, causing an error.