Bug 693179: move gc related variables off C stack
The variables ticks_left and gc_signal were local stack variables in interp()
and gs_call_inter() respectively, a pointer to which was set in the
gs_memory_gc_status_t of each memory "space".

This moves and "merges" the variables into a single entry in gs_lib_ctx, and
has the memory spaces access that variable directly in gs_lib_ctx, rather than
through a pointer belonging to each space.