Fix some error returns in pdfwrite.
Various places in the pdfwrite device we call:

  gs_note_error(x)

and then carry on. The compiler warnings report whinges about
this.

We could fix it by doing:

  (void)gs_note_error(x);

to simply shut the compiler warnings up, but it might be nicer to
do it properly by changing the functions to correctly return
errors.

I'm going to try the latter. Ken can kick me later.