Bug 696499: avoid infinite recursion of Show calls
In the PDF interpreter we keep a procedure called "/Show" in the PDF
graphics state dictionary, which gets changed to suit the current text
rendering settings when we enter a text context.

To cope with a previous broken PDF we prevent changes to the graphics state
under certain circumstances (specifically, after a 'W' operator).

In this case, we have a *very* broken PDF that attempts to draw some text
in these circumstances - as the graphics state is is now "locked", we
cannot update /Show and we end up infinitely recursing.

We now special case /Show to prevent this.