Bug 696502: cope with PaintProc with more gsaves than grestores
If a pattern PaintProc executes one or more gsave operations, and fails to
execute a matching number of grestore operations, we can end up trying to
complete any transparency operations and cache the pattern tile with the
wrong gstate in force.

pattern_paint_cleanup() already dealt with that, so add similar code to
pattern_paint_finish() - as the information is already available on the
stack.