Bug 696498: clean up endian configure checks
The custom endian checks in the configure scripts were implemented because the
built-in test in the most common autoconf version at the time was buggy on
several platforms.

Those autoconf problems were addressed some time and several versions ago, and
the newer versions are widely adopted, so we can safely use the built-in
AC_C_BIGENDIAN() test.