Tweak LCMS2 endian setting in configure
The previous configure setting did not work with the latest endian stuff in
lcms2.h when building a universal binary on Mac. This change resolves that.

Also change the order of the types we check when looking for a 64 bit data
type for color indices - put the more commonly successful one first. Again,
this is in support of universal binaries on Mac.

Noticed in passing.....