text extraction - restore the GC macro
Robin asserts that this macro can have no effect and removed it as it
causes a compiler warning.

Checking this properly  will require more time than I have conveniently
at the moment, and I seem to recall that this was required, in the past
at least.

Because this might result in a difficult to track down GC problem, and
was only removed to prevent a compiler warning, I'm restoring the old
code until I can get the time to walk through it properly and determine
if it is genuinely unused, and if it is, whether it should be.....